Prepare queueing
Related to #394.
