+++ internal/manager/server.go
@@ -2,16 +2,13 @@ package manager
 
 import (
 	"context"
-	"crypto/sha1"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"math/rand"
 	"net/http"
 	"regexp"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -19,35 +16,15 @@ import (
 	kubernetesClient "github.com/opendevstack/pipeline/internal/kubernetes"
 	intrepo "github.com/opendevstack/pipeline/internal/repository"
 	tektonClient "github.com/opendevstack/pipeline/internal/tekton"
-	"github.com/opendevstack/pipeline/pkg/bitbucket"
 	"github.com/opendevstack/pipeline/pkg/config"
+	"github.com/opendevstack/pipeline/pkg/logging"
 	tekton "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 )
 
 const (
 	// allowedChangeRefType is the Bitbucket change ref handled by this service.
 	allowedChangeRefType = "BRANCH"
-	// Label prefix to use for labels applied by this service.
-	labelPrefix = "pipeline.opendevstack.org/"
-	// Label specifying the Bitbucket repository related to the pipeline.
-	repositoryLabel = labelPrefix + "repository"
-	// Label specifying the Git ref (e.g. branch) related to the pipeline.
-	gitRefLabel = labelPrefix + "git-ref"
-	// Label specifying the target stage of the pipeline.
-	stageLabel = labelPrefix + "stage"
-	// tektonAPIVersion specifies the Tekton API version in use
-	tektonAPIVersion = "tekton.dev/v1beta1"
-	// Annotation to set the storage provisioner for a PVC.
-	storageProvisionerAnnotation = "volume.beta.kubernetes.io/storage-provisioner"
-	// PVC finalizer.
-	pvcProtectionFinalizer = "kubernetes.io/pvc-protection"
-	// sharedWorkspaceName is the name of the workspace shared by all tasks
-	sharedWorkspaceName = "shared-workspace"
 	// letterBytes contains letters to use for random strings.
 	letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
 )
@@ -66,7 +43,8 @@ type Server struct {
 	StorageConfig     StorageConfig
 	BitbucketClient   bitbucketInterface
 	PipelineRunPruner PipelineRunPruner
-	pruneMutex        sync.Mutex
+	Mutex             sync.Mutex
+	Logger            logging.LeveledLoggerInterface
 }
 
 type StorageConfig struct {
@@ -102,6 +80,8 @@ type ServerConfig struct {
 	BitbucketClient bitbucketInterface
 	// PipelineRunPruner is responsible to prune pipeline runs.
 	PipelineRunPruner PipelineRunPruner
+	// Logger is the logger to send logging messages to.
+	Logger logging.LeveledLoggerInterface
 }
 
 type PipelineData struct {
@@ -123,11 +103,6 @@ type PipelineData struct {
 	Comment         string `json:"comment"`
 	PullRequestKey  int    `json:"prKey"`
 	PullRequestBase string `json:"prBase"`
-}
-
-type bitbucketInterface interface {
-	bitbucket.CommitClientInterface
-	bitbucket.RawClientInterface
 }
 
 func init() {
@@ -157,7 +132,10 @@ func NewServer(serverConfig ServerConfig) (*Server, error) {
 	if serverConfig.BitbucketClient == nil {
 		return nil, errors.New("bitbucket client is required")
 	}
-	return &Server{
+	if serverConfig.Logger == nil {
+		serverConfig.Logger = &logging.LeveledLogger{Level: logging.LevelError}
+	}
+	s := &Server{
 		KubernetesClient:  serverConfig.KubernetesClient,
 		TektonClient:      serverConfig.TektonClient,
 		BitbucketClient:   serverConfig.BitbucketClient,
@@ -170,39 +148,9 @@ func NewServer(serverConfig ServerConfig) (*Server, error) {
 		TaskSuffix:        serverConfig.TaskSuffix,
 		StorageConfig:     serverConfig.StorageConfig,
 		PipelineRunPruner: serverConfig.PipelineRunPruner,
-	}, nil
-}
-
-type repository struct {
-	Project struct {
-		Key string `json:"key"`
-	} `json:"project"`
-	Slug string `json:"slug"`
-}
-type requestBitbucket struct {
-	EventKey   string     `json:"eventKey"`
-	Repository repository `json:"repository"`
-	Changes    []struct {
-		Type string `json:"type"`
-		Ref  struct {
-			ID        string `json:"id"`
-			DisplayID string `json:"displayId"`
-			Type      string `json:"type"`
-		} `json:"ref"`
-		FromHash string `json:"fromHash"`
-		ToHash   string `json:"toHash"`
-	} `json:"changes"`
-	PullRequest *struct {
-		FromRef struct {
-			Repository   repository `json:"repository"`
-			ID           string     `json:"id"`
-			DisplayID    string     `json:"displayId"`
-			LatestCommit string     `json:"latestCommit"`
-		} `json:"fromRef"`
-	} `json:"pullRequest"`
-	Comment *struct {
-		Text string `json:"text"`
-	} `json:"comment"`
+		Logger:            serverConfig.Logger,
+	}
+	return s, nil
 }
 
 // HandleRoot handles all requests to this service. It performs the following:
@@ -212,18 +160,18 @@ type requestBitbucket struct {
 func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 
 	requestID := randStringBytes(6)
-	log.Println(requestID, "---START---")
+	s.Logger.Infof(requestID, "---START---")
 
 	body, err := ioutil.ReadAll(r.Body)
 	if err != nil {
-		log.Println(requestID, err.Error())
+		s.Logger.Errorf(requestID, err.Error())
 		http.Error(w, "could not read body", http.StatusInternalServerError)
 		return
 	}
 
 	if err := validatePayload(r.Header, body, []byte(s.WebhookSecret)); err != nil {
 		msg := "failed to validate incoming request"
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 		http.Error(w, msg, http.StatusBadRequest)
 		return
 	}
@@ -231,7 +179,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 	req := &requestBitbucket{}
 	if err := json.Unmarshal(body, &req); err != nil {
 		msg := fmt.Sprintf("cannot parse JSON: %s", err)
-		log.Println(requestID, msg)
+		s.Logger.Errorf(requestID, msg)
 		http.Error(w, msg, http.StatusBadRequest)
 		return
 	}
@@ -260,7 +208,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 				change.Ref.Type,
 				allowedChangeRefType,
 			)
-			log.Println(requestID, msg)
+			s.Logger.Warnf(requestID, msg)
 			// According to MDN (https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/418),
 			// "some websites use this response for requests they do not wish to handle [...]".
 			http.Error(w, msg, http.StatusTeapot)
@@ -282,7 +230,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		commitSHA = req.PullRequest.FromRef.LatestCommit
 	} else {
 		msg := fmt.Sprintf("Unsupported event key: %s", req.EventKey)
-		log.Println(requestID, msg)
+		s.Logger.Warnf(requestID, msg)
 		http.Error(w, msg, http.StatusBadRequest)
 		return
 	}
@@ -318,7 +266,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		csha, err := getCommitSHA(s.BitbucketClient, pData.Project, pData.Repository, pData.GitFullRef)
 		if err != nil {
 			msg := "could not get commit SHA"
-			log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+			s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 			http.Error(w, msg, http.StatusInternalServerError)
 			return
 		}
@@ -329,7 +277,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 	skip := shouldSkip(s.BitbucketClient, pData.Project, pData.Repository, commitSHA)
 	if skip {
 		msg := "Commit should be skipped"
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Infof(requestID, fmt.Sprintf("%s: %s", msg, err))
 		// According to MDN (https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/418),
 		// "some websites use this response for requests they do not wish to handle [..]".
 		http.Error(w, msg, http.StatusTeapot)
@@ -338,7 +286,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 	prInfo, err := extractPullRequestInfo(s.BitbucketClient, pData.Project, pData.Repository, commitSHA)
 	if err != nil {
 		msg := "Could not extract PR info"
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 		http.Error(w, msg, http.StatusInternalServerError)
 		return
 	}
@@ -354,7 +302,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 
 	if err != nil {
 		msg := fmt.Sprintf("could not download ODS config for repo %s", pData.Repository)
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 		http.Error(w, msg, http.StatusInternalServerError)
 		return
 	}
@@ -365,7 +313,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		env, err := odsConfig.Environment(pData.Environment)
 		if err != nil {
 			msg := fmt.Sprintf("environment misconfiguration: %s", err)
-			log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+			s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 			http.Error(w, msg, http.StatusInternalServerError)
 			return
 		}
@@ -380,7 +328,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		_, err := s.TektonClient.CreatePipeline(r.Context(), newPipeline, metav1.CreateOptions{})
 		if err != nil {
 			msg := fmt.Sprintf("cannot create pipeline %s", newPipeline.Name)
-			log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+			s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 			http.Error(w, msg, http.StatusInternalServerError)
 			return
 		}
@@ -389,7 +337,7 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		_, err := s.TektonClient.UpdatePipeline(r.Context(), newPipeline, metav1.UpdateOptions{})
 		if err != nil {
 			msg := fmt.Sprintf("cannot update pipeline %s", newPipeline.Name)
-			log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+			s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 			http.Error(w, msg, http.StatusInternalServerError)
 			return
 		}
@@ -399,32 +347,34 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 	err = s.createPVCIfRequired(r.Context(), pData)
 	if err != nil {
 		msg := "cannot create workspace PVC"
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 		http.Error(w, msg, http.StatusInternalServerError)
 		return
 	}
 
+	// Aquire lock to avoid calls timewise close to one another to lead to
+	// parallel pipeline runs.
+	s.Mutex.Lock()
+	defer s.Mutex.Unlock()
+	s.Logger.Infof(requestID, fmt.Sprintf("Starting pruning of pipeline runs related to repository %s ...", pData.Repository))
+	ctxt, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	defer cancel()
+	pipelineRuns, err := listPipelineRuns(s.TektonClient, ctxt, pData.Repository)
+	if err != nil {
+		msg := "could not retrieve existing pipeline runs"
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
+		http.Error(w, msg, http.StatusInternalServerError)
+		return
+	}
+	s.Logger.Debugf(requestID, fmt.Sprintf("Found %d pipeline runs related to repository %s.", len(pipelineRuns.Items), pData.Repository))
+
 	if s.PipelineRunPruner != nil {
 		go func() {
-			// Make sure we do not clean up in parallel, which may lead to
-			// errors or weird results.
-			s.pruneMutex.Lock()
-			defer s.pruneMutex.Unlock()
-			log.Println(requestID, fmt.Sprintf("Starting pruning of pipeline runs related to repository %s ...", pData.Repository))
 			ctxt, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
 			defer cancel()
-			labelMap := map[string]string{repositoryLabel: pData.Repository}
-			pipelineRuns, err := s.TektonClient.ListPipelineRuns(
-				ctxt, metav1.ListOptions{LabelSelector: labels.Set(labelMap).String()},
-			)
-			if err != nil {
-				log.Printf("Could not retrieve existing pipeline runs: %s\n", err)
-				return
-			}
-			log.Println(requestID, fmt.Sprintf("Found %d pipeline runs related to repository %s.", len(pipelineRuns.Items), pData.Repository))
 			err = s.PipelineRunPruner.Prune(ctxt, pipelineRuns.Items)
 			if err != nil {
-				log.Println(fmt.Sprintf(
+				s.Logger.Warnf(fmt.Sprintf(
 					"Pruning pipeline runs of repository %s failed: %s",
 					pData.Repository, err,
 				))
@@ -432,92 +382,20 @@ func (s *Server) HandleRoot(w http.ResponseWriter, r *http.Request) {
 		}()
 	}
 
-	log.Println(requestID, fmt.Sprintf("%+v", pData))
+	s.Logger.Infof(requestID, fmt.Sprintf("%+v", pData))
 
-	_, err = s.createPipelineRun(r.Context(), pData)
+	_, err = createPipelineRun(s.TektonClient, r.Context(), pData)
 	if err != nil {
 		msg := "cannot create pipeline run"
-		log.Println(requestID, fmt.Sprintf("%s: %s", msg, err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("%s: %s", msg, err))
 		http.Error(w, msg, http.StatusInternalServerError)
 		return
 	}
 
 	err = json.NewEncoder(w).Encode(pData)
 	if err != nil {
-		log.Println(requestID, fmt.Sprintf("cannot write body: %s", err))
+		s.Logger.Errorf(requestID, fmt.Sprintf("cannot write body: %s", err))
 		return
-	}
-}
-
-// createPipelineRun creates a PipelineRun resource
-func (s *Server) createPipelineRun(ctxt context.Context, pData PipelineData) (*tekton.PipelineRun, error) {
-	return s.TektonClient.CreatePipelineRun(ctxt, &tekton.PipelineRun{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: fmt.Sprintf("%s-", pData.Name),
-			Labels:       pipelineLabels(pData),
-		},
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: tektonAPIVersion,
-			Kind:       "PipelineRun",
-		},
-		Spec: tekton.PipelineRunSpec{
-			PipelineRef:        &tekton.PipelineRef{Name: pData.Name},
-			ServiceAccountName: "pipeline", // TODO
-			Workspaces: []tekton.WorkspaceBinding{
-				{
-					Name: sharedWorkspaceName,
-					PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-						ClaimName: pData.PVC,
-					},
-				},
-			},
-		},
-	}, metav1.CreateOptions{})
-}
-
-// createPVCIfRequired if it does not exist yet
-func (s *Server) createPVCIfRequired(ctxt context.Context, pData PipelineData) error {
-	_, err := s.KubernetesClient.GetPersistentVolumeClaim(ctxt, pData.PVC, metav1.GetOptions{})
-	if err != nil {
-		if !kerrors.IsNotFound(err) {
-			return fmt.Errorf("could not determine if %s already exists: %w", pData.PVC, err)
-		}
-		vm := corev1.PersistentVolumeFilesystem
-		pvc := &corev1.PersistentVolumeClaim{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:        pData.PVC,
-				Labels:      map[string]string{repositoryLabel: pData.Repository},
-				Finalizers:  []string{pvcProtectionFinalizer},
-				Annotations: map[string]string{},
-			},
-			Spec: corev1.PersistentVolumeClaimSpec{
-				AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-				Resources: corev1.ResourceRequirements{
-					Requests: corev1.ResourceList{
-						corev1.ResourceName(corev1.ResourceStorage): resource.MustParse(s.StorageConfig.Size),
-					},
-				},
-				StorageClassName: &s.StorageConfig.ClassName,
-				VolumeMode:       &vm,
-			},
-		}
-		if s.StorageConfig.Provisioner != "" {
-			pvc.Annotations[storageProvisionerAnnotation] = s.StorageConfig.Provisioner
-		}
-		_, err := s.KubernetesClient.CreatePersistentVolumeClaim(ctxt, pvc, metav1.CreateOptions{})
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// pipelineLabels returns a map of labels to apply to pipelines and related runs.
-func pipelineLabels(data PipelineData) map[string]string {
-	return map[string]string{
-		repositoryLabel: data.Repository,
-		gitRefLabel:     makeValidLabelValue("", data.GitRef, 63),
-		stageLabel:      data.Stage,
 	}
 }
 
@@ -546,300 +424,12 @@ func mappingBranchMatch(mappingBranch, testBranch string) bool {
 	return false
 }
 
-func getCommitSHA(bitbucketClient bitbucket.CommitClientInterface, project, repository, gitFullRef string) (string, error) {
-	commitList, err := bitbucketClient.CommitList(project, repository, bitbucket.CommitListParams{
-		Until: gitFullRef,
-	})
-	if err != nil {
-		return "", fmt.Errorf("could not get commit list: %w", err)
-	}
-	return commitList.Values[0].ID, nil
-}
-
 func determineProject(serverProject string, projectParam string) string {
 	projectParam = strings.ToLower(projectParam)
 	if len(projectParam) > 0 {
 		return projectParam
 	}
 	return strings.ToLower(serverProject)
-}
-
-// makePipelineName generates the name of the pipeline.
-// According to the Kubernetes label rules, a maximum of 63 characters is
-// allowed, see https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#syntax-and-character-set.
-// Therefore, the name might be truncated. As this could cause potential clashes
-// between similar named branches, we put a short part of the branch hash value
-// into the name name to make this very unlikely.
-// We cut the pipeline name at 55 chars to allow e.g. pipeline runs to add suffixes.
-func makePipelineName(component string, branch string) string {
-	// 55 is derived from K8s label max length minus room for generateName suffix.
-	return makeValidLabelValue(component+"-", branch, 55)
-}
-
-func makeValidLabelValue(prefix, branch string, maxLength int) string {
-	// Cut all non-alphanumeric characters
-	safeCharsRegex := regexp.MustCompile("[^-a-zA-Z0-9]+")
-	result := prefix + safeCharsRegex.ReplaceAllString(
-		strings.Replace(branch, "/", "-", -1),
-		"",
-	)
-	result = fitStringToMaxLength(result, maxLength)
-	return strings.ToLower(result)
-}
-
-func makePVCName(component string) string {
-	pvcName := fmt.Sprintf("ods-workspace-%s", strings.ToLower(component))
-	return fitStringToMaxLength(pvcName, 63) // K8s label max length to be on the safe side.
-}
-
-// fitStringToMaxLength ensures s is not longer than max.
-// If s is longer than max, it shortenes s and appends a unique, consistent
-// suffix so that multiple invocations produce the same result. The length
-// of the shortened string will be equal to max.
-func fitStringToMaxLength(s string, max int) string {
-	if len(s) <= max {
-		return s
-	}
-	suffixLength := 7
-	shortened := s[0 : max-suffixLength-1]
-	h := sha1.New()
-	_, err := h.Write([]byte(s))
-	if err != nil {
-		return shortened
-	}
-	bs := h.Sum(nil)
-	suffix := fmt.Sprintf("%x", bs)
-	return fmt.Sprintf("%s-%s", shortened, suffix[0:suffixLength])
-}
-
-func assemblePipeline(odsConfig *config.ODS, data PipelineData, taskKind tekton.TaskKind, taskSuffix string) *tekton.Pipeline {
-
-	var tasks []tekton.PipelineTask
-	tasks = append(tasks, tekton.PipelineTask{
-		Name:    "ods-start",
-		TaskRef: &tekton.TaskRef{Kind: taskKind, Name: "ods-start" + taskSuffix},
-		Workspaces: []tekton.WorkspacePipelineTaskBinding{
-			{Name: "source", Workspace: sharedWorkspaceName},
-		},
-		Params: []tekton.Param{
-			{
-				Name: "url",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.git-repo-url)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "git-full-ref",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.git-full-ref)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "project",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.project)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "pr-key",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.pr-key)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "pr-base",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.pr-base)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "pipeline-run-name",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(context.pipelineRun.name)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "environment",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.environment)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "version",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(params.version)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-		},
-	})
-	if len(odsConfig.Pipeline.Tasks) > 0 {
-		odsConfig.Pipeline.Tasks[0].RunAfter = append(odsConfig.Pipeline.Tasks[0].RunAfter, "ods-start")
-		tasks = append(tasks, odsConfig.Pipeline.Tasks...)
-	}
-
-	var finallyTasks []tekton.PipelineTask
-	finallyTasks = append(finallyTasks, odsConfig.Pipeline.Finally...)
-
-	finallyTasks = append(finallyTasks, tekton.PipelineTask{
-		Name:    "ods-finish",
-		TaskRef: &tekton.TaskRef{Kind: taskKind, Name: "ods-finish" + taskSuffix},
-		Workspaces: []tekton.WorkspacePipelineTaskBinding{
-			{Name: "source", Workspace: sharedWorkspaceName},
-		},
-		Params: []tekton.Param{
-			{
-				Name: "pipeline-run-name",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(context.pipelineRun.name)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-			{
-				Name: "aggregate-tasks-status",
-				Value: tekton.ArrayOrString{
-					StringVal: "$(tasks.status)",
-					Type:      tekton.ParamTypeString,
-				},
-			},
-		},
-	})
-
-	p := &tekton.Pipeline{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   data.Name,
-			Labels: pipelineLabels(data),
-		},
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: tektonAPIVersion,
-			Kind:       "Pipeline",
-		},
-		Spec: tekton.PipelineSpec{
-			Description: "ODS",
-			Params: []tekton.ParamSpec{
-				{
-					Name: "repository",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.Repository,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "project",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.Project,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "component",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.Component,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "git-repo-url",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.GitURI,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "git-full-ref",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.GitFullRef,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "pr-key",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: strconv.Itoa(data.PullRequestKey),
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "pr-base",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.PullRequestBase,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "environment",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.Environment,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-				{
-					Name: "version",
-					Type: "string",
-					Default: &tekton.ArrayOrString{
-						StringVal: data.Version,
-						Type:      tekton.ParamTypeString,
-					},
-				},
-			},
-			Tasks: tasks,
-			Workspaces: []tekton.PipelineWorkspaceDeclaration{
-				{
-					Name: sharedWorkspaceName,
-				},
-			},
-			Finally: finallyTasks,
-		},
-	}
-	return p
-}
-
-type prInfo struct {
-	ID   int
-	Base string
-}
-
-func extractPullRequestInfo(bitbucketClient bitbucket.CommitClientInterface, projectKey, repositorySlug, gitCommit string) (prInfo, error) {
-	var i prInfo
-
-	prPage, err := bitbucketClient.CommitPullRequestList(projectKey, repositorySlug, gitCommit)
-	if err != nil {
-		return i, err
-	}
-
-	for _, v := range prPage.Values {
-		if !v.Open {
-			continue
-		}
-		i.ID = v.ID
-		i.Base = v.ToRef.ID
-		break
-	}
-
-	return i, nil
-}
-
-func shouldSkip(bitbucketClient bitbucket.CommitClientInterface, projectKey, repositorySlug, gitCommit string) bool {
-	c, err := bitbucketClient.CommitGet(projectKey, repositorySlug, gitCommit)
-	if err != nil {
-		return false
-	}
-	return isCiSkipInCommitMessage(c.Message)
 }
 
 /** Looks in commit message for the following strings
